Use binding to force array coercion
With current Rakudo (2015.05), the arrays weren't being extracted from the
parsed elements properly (only a single element was being returned) and thus
arrays of rules were turning up where single rules were expected.  This now
allows the script to run to completion, and to extract the same number of
rules, zones and links as Rakudo 2014.01, where this script was found to run.