Fix issue with linked timezones in "backward"
The "backward" file broke the idea of being able to generate and
update per timezone data file, because it mentions linked timezones
that do not actually occur in that timezone file.  And because of
that, all linked timezones in "backward" had disappeared.  So now
we have 594 supported timezone files (119 more than before).